Fix alerts
Fixed a bug where alerts were never reported. Turns out that the "Time Machine" function doesn't have the laerts, so I had to create a second call to pull alert data. If there is an alert, the tweet contains a link to the NOAA alert for my state (Arkansas). You'll need to change that to suit your needs.